Chapter 11 Exercise
To round off the chapter, I followed the instructions to build a website, harnessing all the skills that we had previously gone, from forms, menus and dropdown lists, all the way to CSS. I was instructed to add three html files to my 'blue_developer', upon which I started CSS in the previous exercise, with the names, about, contact and browse, which were already linked to in index. I was provided a page outline that acted as a template, upon which I copy and pasted content from the 'blue web developer - parts' folder, being advised to inspect and understand the HTML I was inserting. I then added to my style.css file for the new elements, classes and tags that were now present, and before I knew it, I had a website, identical to the 'blue web developer - done' folder.