Chapter 10 Exercise
This was focussed on CSS. Because of the nature of the tutorials and resources, I did not use this folder, as everything I need waqs already provided wihtin the previous chapter, which explains why no 'blue_developer' with my work can be seen in this directory. But simply for informational purposes and reference, I felt it was fitting to include this.