Chapter 21 - Graphics
The main focus of this exercise was 'canvas/canvas.js', with 'index.html' already being provided for me, not requiring any amendments. Since javascript is practically the same as java, understanding the code was quite straightforward, it was simply understanding the rpocess and classes used that I was not familiar with.